backend/requirements.txt
# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.5.2

# --- Database ---
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9   ; platform_system != "Windows" or python_version < "3.13"

# --- Transcription (local, CPU) ---
faster-whisper==1.0.3
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.4

# --- LLM local via Ollama ---
ollama==0.3.3
httpx==0.27.2

# --- NLP ---
scikit-learn==1.5.2
sentence-transformers==3.1.1
nltk==3.9.1

# --- Vector store (local) ---
chromadb==0.5.7
faiss-cpu==1.8.0.post1

# --- Speaker diarization (optional, CPU) ---
# pyannote.audio==3.3.1  # opcional, requiere token HF
resemblyzer==0.1.4

# --- ETL / Big Data ---
pandas==2.2.3
pyarrow==17.0.0
# pyspark==3.5.3  # opcional

# --- Async task queue (ligero, local) ---
# redis==5.0.8  # opcional
# celery==5.4.0 # opcional

# --- Export ---
reportlab==4.2.5
python-docx==1.1.2

# --- Utilities ---
loguru==0.7.2
tqdm==4.66.5
python-dotenv==1.0.1
